+++ packages/apptive_grid_core/lib/src/model/data_entity.dart
@@ -93,6 +93,8 @@ abstract class DataEntity<T, S> with FilterableMixin {
         return UriDataEntity.fromJson(json);
       case DataType.email:
         return EmailDataEntity(json);
+      case DataType.phoneNumber:
+        return PhoneNumberDataEntity(json);
     }
   }
 }
@@ -505,6 +507,15 @@ class UriDataEntity extends DataEntity<Uri, String> {
 class EmailDataEntity extends DataEntity<String, String> {
   /// Creates a new StringDataEntity Object with value [value]
   EmailDataEntity([super.value]);
+
+  @override
+  String? get schemaValue => value;
+}
+
+/// [DataEntity] representing an email
+class PhoneNumberDataEntity extends DataEntity<String, String> {
+  /// Creates a new StringDataEntity Object with value [value]
+  PhoneNumberDataEntity([super.value]);
 
   @override
   String? get schemaValue => value;